requirements.txt
fastapi[all]
python-dotenv
requests
beautifulsoup4
botocore
boto3
psycopg2-binary